Modify the Gameboard API to return the placedShips object.
The placedShipsCoords array, which was returned previously, is more
useful for internal tracking of placed ships for preventing placing a
ship over the coordinates of an existing ship.

For the purposes of the GUI, the info in the placedShipsCoords can be
reconstructed from the placedShips object. If it makes sense in the
future, when the GUI part of the application is created, to make the
coordinates array available, it can be made public at that time.