Change "Restart" to "To Ship Placement"
As we don't want to allow people to cheat by playing the game and
finding the opponent ships and then restarting with the same ship
placement, the restart button should go to ship placement screen,
clearing the ship placements from both players. This is done by creating
new players with the same name and isAi parameters.