Remove async delays from the controller and fix restart tests
The delays will be implemented in the UI only - this will make automated
tests run faster and simplify the controller.
The restart tests were failing because the state machine of the game was
changed - at restart, the game goes back to ship placing, rather than
into the same game with the same ship placements.