Disable interactivity when UI is delaying display.
In a single player game, the display delaying is done on computer moves,
in order to simulate the computer thinking (the player should not be
able to shoot during the computer turn).
In a multiplayer game, the display delaying is done when the turn is
changed, before the pass screen is rendered, in order for the player to
see the missed shot (he should not be able to shoot again until the
other player shoots, passes the device back and the pass modal is
clicked away (shot is primed).